requirement.txt
google-generativeai>=0.4.1
python-dotenv>=1.0.0
pyttsx3>=2.90